refactor isPoolNotInitialized to be used in event of no candle data
